Access logs to a configurable IP by socket programming
Created a program where the administrator is listening to the logs
flowing to a configurable IP address. My tool takes this IP address and
port as an input and then listen on the same address. If the input
matches with the existing log, then it gets displayed otherwise it’s
rejected.